// libs/firefly-shared/src/state/document/list/list.actions.enum.ts
export enum ActionsList {
  Reset = '[List] Reset',
  Dirty = '[List] Dirty',
  Get = '[List] Get',
  SetId = '[List] SetId',
  Set = '[List] Set',
  Patch = '[List] Patch',
  PatchMetadata = '[List] Patch Metadata',
  Create = '[List] Create',
  Update = '[List] Update',
  Save = '[List] Save',
  Delete = '[List] Delete',

  SetIdAnonymous = '[List] SetIdAnonymous',

  IconClear = '[List] Icon Clear',
  IconUriSet = '[List] Icon Uri Set',
  IconPathSet = '[List] Icon Path Set',
  IconCreate = '[List] Icon Create',

  ImageClear = '[List] Image Clear',
  ImagesUpdate = '[List] Images Update',
  ImageSet = '[List] Image Set',

  EventsGet = '[List] Events Get',
  EventsGetPending = '[List] Events Get Pending',
  EventsGetAnonymous = '[List] Events Get Anonymous',
  EventsSet = '[List] Events Set',
  EventsAdd = '[List] Events Add'
}